+++ test/support/web/controllers/live_query_controller.ex
@@ -2,7 +2,7 @@ defmodule DrabTestApp.LiveQueryController do
   @moduledoc false
 
   use DrabTestApp.Web, :controller
-  use Drab.Controller
+  # use Drab.Controller
 
   require Logger