+++ lib/drab/channel.ex
@@ -132,6 +132,8 @@ defmodule Drab.Channel do
         },
         socket
       ) do
+    # IO.inspect payload
+    socket = Phoenix.Socket.assign(socket, :__sender_drab_commander_id, payload["drab_commander_id"])
     verify_and_cast(:event, [payload, event_handler_function, reply_to], socket)
   end