+++ test/support/web/commanders/shared1_commander.ex
@@ -5,7 +5,10 @@ defmodule DrabTestApp.Shared1Commander do
   # onload(:page_loaded)
   public(:button_clicked)
 
-  def button_clicked(_socket, _) do
+  def button_clicked(socket, sender) do
+    IO.inspect socket
+    set_prop socket, this_commander(sender) <> " .spaceholder1", innerText: "changed"
+    poke socket, text: "changed in commander"
   end
 
   # def page_loaded(socket) do