+++ test/support/web/controllers/share_controller.ex
@@ -5,6 +5,6 @@ defmodule DrabTestApp.ShareController do
   use Drab.Controller, commanders: [DrabTestApp.Shared1Commander, DrabTestApp.Shared2Commander]
 
   def index(conn, _params) do
-    render(conn, "index.html")
+    render(conn, "index.html", text: "assigned in controller")
   end
 end