+++ lib/drab/channel.ex
@@ -133,8 +133,11 @@ defmodule Drab.Channel do
         },
         socket
       ) do
-    # IO.inspect payload
-    socket = Phoenix.Socket.assign(socket, :__sender_drab_commander_id, payload["drab_commander_id"])
+    socket = if payload["drab_commander_id"] do
+      Phoenix.Socket.assign(socket, :__sender_drab_commander_id, payload["drab_commander_id"])
+    else
+      socket
+    end
     verify_and_cast(:event, [payload, event_handler_function, reply_to], socket)
   end